+++ chainflip-insights/Feeders/Funding/FundingResponse.cs
@@ -50,7 +50,7 @@ namespace ChainflipInsights.Feeders.Funding
         public double Amount { get; set; }
         
         [JsonPropertyName("epochId")] 
-        public double Epoch { get; set; }
+        public ulong Epoch { get; set; }
         
         [JsonPropertyName("accountByAccountId")] 
         public FundingInfoAccount Account { get; set; }
